chore: add upload_to for bulk ops csv_files, move tasks tests to test_tasks
