+++ course_discovery/apps/course_metadata/models.py
@@ -5032,6 +5032,7 @@ class BulkOperationTask(TimeStampedModel):
     Model to store information related to bulk operations.
     """
     csv_file = models.FileField(
+        upload_to=UploadToFieldNamePath(populate_from='created', path='bulk_operations'),
         validators=[FileExtensionValidator(allowed_extensions=['csv'])],
         help_text=_("It expects the data will be provided in a csv file format "
                     "with first row containing all the headers."),
